Create # Start Outlook in Safe Mode and Disable.ps1
Here’s a PowerShell script that you can use to start Outlook in safe mode and disable add-ins.  Save this script with a .ps1 extension and execute it using a PowerShell environment. Make sure to update the $OutlookPath variable with the correct path to your Outlook executable if it's not in the default location.

Note: Running this script might require administrative privileges on your system, and the success of this script depends on your Outlook installation and system configuration.